added buffer loop and converted processing to f32

+++ code/src/audio_processing/audio_modules/invert.rs
@@ -2,7 +2,7 @@
 struct Invert;
 
 impl Invert {
-    fn process(&self, buffer: &mut [i16]) {
+    fn process(&self, buffer: &mut [f32]) {
         for sample in buffer.iter_mut() {
             *sample = *sample * -1;
         }
